Correct type of weight in item class

+++ game_files/basic_classes/item.cpp
@@ -7,7 +7,7 @@ string Item::get_name() {
     return name;
 }
 
-float Item::get_weight() {
+int Item::get_weight() {
     return weight;
 }
