+++ game_files/basic_classes/item.hpp
@@ -5,11 +5,11 @@ using namespace std;
 class Item {
 private:
     string name;
-    float weight;
+    int weight;
 
 public:
     string get_name();
-    float get_weight();
+    int get_weight();
     virtual void execute();
     virtual ~Item();
 };
